+++ extension/main.js
@@ -3,32 +3,58 @@ const deminifier = new Deminifier(mapping);
 const crateSearcher = new CrateSearch(crateIndex);
 const attributeSearcher = new AttributeSearch();
 const bookSearcher = new BookSearch(booksIndex);
-
+const lintSearcher = new LintSearch(lintsIndex);
+const stdSearcher = new StdSearch(searchIndex);
+const crateDocSearchManager = new CrateDocSearchManager();
 const commandManager = new CommandManager();
 
 const defaultSuggestion = `Search std ${c.match("docs")}, ${c.match("crates")} (!), builtin ${c.match("attributes")} (#), official ${c.match("books")} (%), and ${c.match("error codes")}, etc in your address bar instantly!`;
 const omnibox = new Omnibox(c.browser, defaultSuggestion, c.isChrome ? 8 : 6);
 
+let formatDoc = (index, doc) => {
+    let description = doc.displayPath + c.match(doc.name);
+    if (doc.desc) {
+        description += " - " + c.dim(c.escape(doc.desc));
+    }
+    return {content: doc.href, description};
+};
+
 omnibox.bootstrap({
     onSearch: (query) => {
-        return window.search(query);
+        return stdSearcher.search(query);
     },
-    onFormat: (index, doc) => {
-        let description = doc.displayPath + c.match(doc.name);
-        if (doc.desc) {
-            description += " - " + c.dim(c.escape(doc.desc));
-        }
-        return {content: doc.href, description};
-    },
+    onFormat: formatDoc,
     onAppend: (query) => {
         return [{
-            content: `${window.rootPath}std/index.html?search=` + encodeURIComponent(query),
-            description: `Search Rust docs ${ c.match(query) } on ${ settings.isOfflineMode ? "offline mode" : "https://doc.rust-lang.org"}`,
-        }]
+            content: stdSearcher.getSearchUrl(query),
+            description: `Search Rust docs ${c.match(query)} on ${settings.isOfflineMode ? "offline mode" : stdSearcher.rootPath}`,
+        }];
     },
     onSelected: (query, result) => {
         HistoryCommand.record(query, result);
     }
+});
+
+omnibox.addPrefixQueryEvent("@", {
+    onSearch: (query) => {
+        return crateDocSearchManager.search(query);
+    },
+    onFormat: (index, item) => {
+        if (item.hasOwnProperty("content")) {
+            // 1. Crate list header.
+            // 2. Crate result footer
+            return item;
+        } else if (item.hasOwnProperty("href")) {
+            return formatDoc(index, item);
+        } else {
+            // Crate name list.
+            let content = `@${item.name}`;
+            return {
+                content,
+                description: `${c.match(content)} - ${c.dim(item.doc)}`,
+            }
+        }
+    },
 });
 
 omnibox.addPrefixQueryEvent("!", {
@@ -36,8 +62,6 @@ omnibox.addPrefixQueryEvent("!", {
     searchPriority: 1,
     onSearch: (query) => {
         this.docMode = query.startsWith("!!");
-        this.rawQuery = query.replace(/[!\s]/g, "");
-        query = this.rawQuery.replace(/[-_]*/ig, "");
         return crateSearcher.search(query);
     },
     onFormat: (index, crate) => {
@@ -46,10 +70,11 @@ omnibox.addPrefixQueryEvent("!", {
             description: `${this.docMode ? "Docs" : "Crate"}: ${c.match(crate.id)} v${crate.version} - ${c.dim(c.escape(crate.description))}`,
         };
     },
-    onAppend: () => {
+    onAppend: (query) => {
+        let keyword = query.replace(/[!\s]/g, "");
         return [{
-            content: "https://crates.io/search?q=" + encodeURIComponent(this.rawQuery),
-            description: "Search Rust crates for " + c.match(this.rawQuery) + " on https://crates.io",
+            content: "https://crates.io/search?q=" + encodeURIComponent(keyword),
+            description: "Search Rust crates for " + c.match(keyword) + " on https://crates.io",
         }, {
             content: "remind",
             description: `Remind: ${c.dim("We only indexed the top 20K crates. Sorry for the inconvenience if your desired crate not show.")}`,
@@ -90,12 +115,6 @@ omnibox.addRegexQueryEvent(/e\d{2,4}$/i, {
     }
 });
 
-omnibox.addPrefixQueryEvent(":", {
-    onSearch: (query) => {
-        return commandManager.execute(query);
-    },
-});
-
 omnibox.addPrefixQueryEvent("%", {
     onSearch: (query) => {
         return bookSearcher.search(query);
@@ -104,9 +123,28 @@ omnibox.addPrefixQueryEvent("%", {
         let parentTitles = page.parentTitles || [];
         return {
             content: page.url,
-            description: `${ [...parentTitles.map(t => c.escape(t)), c.match(c.escape(page.title))].join(" > ") } - ${c.dim(page.name)}`
+            description: `${[...parentTitles.map(t => c.escape(t)), c.match(c.escape(page.title))].join(" > ")} - ${c.dim(page.name)}`
         }
     }
+});
+
+const LINT_URL = "https://rust-lang.github.io/rust-clippy/master/";
+omnibox.addPrefixQueryEvent(">", {
+    onSearch: (query) => {
+        return lintSearcher.search(query);
+    },
+    onFormat: (index, lint) => {
+        return {
+            content: `${LINT_URL}#${lint.name}`,
+            description: `Clippy lint: [${lint.level}] ${c.match(lint.name)} - ${c.dim(c.escape(lint.description))}`,
+        }
+    },
+});
+
+omnibox.addPrefixQueryEvent(":", {
+    onSearch: (query) => {
+        return commandManager.execute(query);
+    },
 });
 
 window.crateSearcher = crateSearcher;